# swagger_output.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "ATS Documentation",
        "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
    },
    "host": "localhost:3000",
    "basePath": "/",
    "schemes": ["http"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
...
        "edit/{id}": {
            "patch": {
                "description": "",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {}
            }
        },
...
    },
...
}